OptimalTrainControl

Documentation for OptimalTrainControl.

OptimalTrainControl.altitude

OptimalTrainControl.gradient

OptimalTrainControl.segmentize!

OptimalTrainControl.speedlimit

OptimalTrainControl.altitude — Method

altitude(track, position)

Return altitude (in metres) at position on track.

OptimalTrainControl.gradient — Method

gradient(track, position)

Return gradient (in rise over run) at position on track. For example, a gradient of a segment which rises 10 metres over 100 metres of distance has gradient of 0.1.

OptimalTrainControl.segmentize! — Method

segmentize!(track)

Modify track.x_segments such that its elements mark starts of track parts on which both gradient and speed limit are constant.

OptimalTrainControl.speedlimit — Method

speedlimit(track, position)

Return speedlimit (in metres per second) at position on track.